Implement abstract ParallelIndexer class
