+++ Model/Indexer/Data/ModeSwitcher.php
@@ -41,7 +41,7 @@ use Magento\Framework\Search\Request\Dimension;
 use Magento\Store\Model\Indexer\WebsiteDimensionProvider;
 use Magento\Customer\Model\Indexer\CustomerGroupDimensionProvider;
 use Magento\Indexer\Model\DimensionMode;
-use Magento\Indexer\Model\ModeSwitcherInterface;
+use Nosto\Tagging\Model\Indexer\Dimensions\ModeSwitcherInterface;
 use Nosto\Tagging\Model\Indexer\Data\DimensionModeConfiguration;
 use Nosto\Tagging\Model\Indexer\Data\ModeSwitcherConfiguration;
 
@@ -94,7 +94,7 @@ class ModeSwitcher implements ModeSwitcherInterface
     /**
      * @return string
      */
-    public function getMode()
+    public function getMode(): string
     {
         return $this->modeSwitcherConfiguration->getMode();
     }
