+++ etc/di.xml
@@ -140,11 +140,4 @@
             </argument>
         </arguments>
     </type>
-    <type name="Nosto\Tagging\Model\Indexer\Data">
-        <arguments>
-            <argument name="dimensionProvider" xsi:type="object" shared="false">
-                Magento\Store\Model\StoreDimensionProvider
-            </argument>
-        </arguments>
-    </type>
 </config>
